zrobić debugowanie findroots

+++ PolyRoots/src/Config.h
@@ -2,4 +2,4 @@
 
 const double PRECISION = 0.05;  // 0.5/0.005
 const int PRECISION_NEWTON = 2;	 // 2-3
-const int DELTA = 1e-3;
+const double DELTA = 1e-3;
